handle arrays, run on common and std msgs

+++ turtlewatch/packages/msg-gen/src/msg-out/Vector3Message.py
@@ -1,9 +1,8 @@
 from dataclasses import dataclass
 
+
 @dataclass
 class Vector3Message:
-    x: int
-    y: int
-    z: int
-
-    
+    x: float
+    y: float
+    z: float
